refactor tf upgrader
