+++ cli/internal/upgrade/upgrade.go
@@ -1,0 +1,10 @@
+/*
+Copyright (c) Edgeless Systems GmbH
+
+SPDX-License-Identifier: AGPL-3.0-only
+*/
+
+/*
+Package upgrade provides functionality to upgrade the cluster and it's resources
+*/
+package upgrade
